Release v0.333.0
This commit is the built version of Nuclide suitable for apm and npm.

+++ pkg/nuclide-hg-repository/lib/HgRepositoryProvider.js
@@ -1,3 +1,72 @@
+"use strict";
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.default = void 0;
+
+function _UniversalDisposable() {
+  const data = _interopRequireDefault(require("../../../modules/nuclide-commons/UniversalDisposable"));
+
+  _UniversalDisposable = function () {
+    return data;
+  };
+
+  return data;
+}
+
+function _nuclideAnalytics() {
+  const data = require("../../nuclide-analytics");
+
+  _nuclideAnalytics = function () {
+    return data;
+  };
+
+  return data;
+}
+
+function _nuclideRemoteConnection() {
+  const data = require("../../nuclide-remote-connection");
+
+  _nuclideRemoteConnection = function () {
+    return data;
+  };
+
+  return data;
+}
+
+function _nuclideHgRepositoryClient() {
+  const data = require("../../nuclide-hg-repository-client");
+
+  _nuclideHgRepositoryClient = function () {
+    return data;
+  };
+
+  return data;
+}
+
+function _log4js() {
+  const data = require("log4js");
+
+  _log4js = function () {
+    return data;
+  };
+
+  return data;
+}
+
+function _nuclideSourceControlHelpers() {
+  const data = require("../../nuclide-source-control-helpers");
+
+  _nuclideSourceControlHelpers = function () {
+    return data;
+  };
+
+  return data;
+}
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
 /**
  * Copyright (c) 2015-present, Facebook, Inc.
  * All rights reserved.
@@ -5,29 +74,10 @@
  * This source code is licensed under the license found in the LICENSE file in
  * the root directory of this source tree.
  *
- * @flow strict-local
+ *  strict-local
  * @format
  */
-
-import type {NuclideUri} from 'nuclide-commons/nuclideUri';
-
-import UniversalDisposable from 'nuclide-commons/UniversalDisposable';
-import {trackTiming} from '../../nuclide-analytics';
-import {
-  RemoteDirectory,
-  getHgServiceByNuclideUri,
-} from '../../nuclide-remote-connection';
-import {HgRepositoryClient} from '../../nuclide-hg-repository-client';
-import {getLogger} from 'log4js';
-import {findHgRepository} from '../../nuclide-source-control-helpers';
-import invariant from 'assert';
-
-const logger = getLogger('nuclide-hg-repository');
-
-type RefCountedRepo = {
-  refCount: number,
-  repo: HgRepositoryClient,
-};
+const logger = (0, _log4js().getLogger)('nuclide-hg-repository');
 
 /**
  * @param directory Either a RemoteDirectory or Directory we are interested in.
@@ -41,64 +91,62 @@ type RefCountedRepo = {
  *    repository (i.e. if it's a remote directory, the URI minus the hostname).
  *  If the directory is not part of a Mercurial repository, returns null.
  */
-function getRepositoryDescription(
-  directory: atom$Directory | RemoteDirectory,
-): ?{
-  originURL: ?string,
-  repoPath: string,
-  workingDirectoryPath: NuclideUri,
-} {
-  if (directory instanceof RemoteDirectory) {
+function getRepositoryDescription(directory) {
+  if (directory instanceof _nuclideRemoteConnection().RemoteDirectory) {
     const repositoryDescription = directory.getHgRepositoryDescription();
-    if (
-      repositoryDescription == null ||
-      repositoryDescription.repoPath == null
-    ) {
+
+    if (repositoryDescription == null || repositoryDescription.repoPath == null) {
       return null;
     }
+
     const serverConnection = directory._server;
-    const {repoPath, originURL, workingDirectoryPath} = repositoryDescription;
-    // These paths are all relative to the remote fs. We need to turn these into URIs.
+    const {
+      repoPath,
+      originURL,
+      workingDirectoryPath
+    } = repositoryDescription; // These paths are all relative to the remote fs. We need to turn these into URIs.
+
     const repoUri = serverConnection.getUriOfRemotePath(repoPath);
-    const workingDirectoryUri = serverConnection.getUriOfRemotePath(
-      workingDirectoryPath,
-    );
+    const workingDirectoryUri = serverConnection.getUriOfRemotePath(workingDirectoryPath);
     return {
       originURL,
       repoPath: repoUri,
-      workingDirectoryPath: workingDirectoryUri,
+      workingDirectoryPath: workingDirectoryUri
     };
   } else {
-    const repositoryDescription = findHgRepository(directory.getPath());
+    const repositoryDescription = (0, _nuclideSourceControlHelpers().findHgRepository)(directory.getPath());
+
     if (repositoryDescription == null) {
       return null;
     }
-    const {repoPath, originURL, workingDirectoryPath} = repositoryDescription;
+
+    const {
+      repoPath,
+      originURL,
+      workingDirectoryPath
+    } = repositoryDescription;
     return {
       originURL,
       repoPath,
-      workingDirectoryPath,
+      workingDirectoryPath
     };
   }
 }
 
-export default class HgRepositoryProvider {
-  // Allow having multiple project roots under the same repo while sharing
-  // the underlying HgRepositoryClient.
-  _activeRepositoryClients: Map<string, RefCountedRepo> = new Map();
+class HgRepositoryProvider {
+  constructor() {
+    this._activeRepositoryClients = new Map();
+  }
 
-  repositoryForDirectory(
-    directory: atom$Directory | RemoteDirectory,
-  ): Promise<?HgRepositoryClient> {
+  repositoryForDirectory(directory) {
     return Promise.resolve(this.repositoryForDirectorySync(directory));
   }
 
-  repositoryForDirectorySync(
-    directory: atom$Directory | RemoteDirectory,
-  ): ?HgRepositoryClient {
-    return trackTiming('hg-repository.repositoryForDirectorySync', () => {
+  repositoryForDirectorySync(directory) {
+    return (0, _nuclideAnalytics().trackTiming)('hg-repository.repositoryForDirectorySync', () => {
       try {
         const repositoryDescription = getRepositoryDescription(directory);
+
         if (!repositoryDescription) {
           return null;
         }
@@ -106,77 +154,76 @@ export default class HgRepositoryProvider {
         const {
           originURL,
           repoPath,
-          workingDirectoryPath,
-        } = repositoryDescription;
-
-        // extend the underlying instance of HgRepositoryClient to prevent
+          workingDirectoryPath
+        } = repositoryDescription; // extend the underlying instance of HgRepositoryClient to prevent
         // having multiple clients for multiple project roots inside the same
         // repository folder
+
         const activeRepositoryClients = this._activeRepositoryClients;
         let activeRepoClientInfo = activeRepositoryClients.get(repoPath);
 
         if (activeRepoClientInfo != null) {
           activeRepoClientInfo.refCount++;
         } else {
-          const hgService = getHgServiceByNuclideUri(workingDirectoryPath);
-          const activeRepoClient = new HgRepositoryClient(repoPath, hgService, {
+          const hgService = (0, _nuclideRemoteConnection().getHgServiceByNuclideUri)(workingDirectoryPath);
+          const activeRepoClient = new (_nuclideHgRepositoryClient().HgRepositoryClient)(repoPath, hgService, {
             workingDirectoryPath,
-            originURL,
+            originURL
           });
-
           activeRepoClientInfo = {
             refCount: 1,
-            repo: activeRepoClient,
+            repo: activeRepoClient
           };
           activeRepositoryClients.set(repoPath, activeRepoClientInfo);
         }
 
         let destroyed = false;
-
-        const localDisposables = new UniversalDisposable();
-
+        const localDisposables = new (_UniversalDisposable().default)();
         /* eslint-disable no-inner-declarations */
+
         function ProjectHgRepositoryClient() {
-          this.getProjectDirectory = function(): NuclideUri {
+          this.getProjectDirectory = function () {
             return directory.getPath();
           };
 
-          this.destroy = function(): void {
-            invariant(activeRepoClientInfo != null);
+          this.destroy = function () {
+            if (!(activeRepoClientInfo != null)) {
+              throw new Error("Invariant violation: \"activeRepoClientInfo != null\"");
+            }
+
             if (!destroyed && --activeRepoClientInfo.refCount === 0) {
               destroyed = true;
               activeRepoClientInfo.repo.destroy();
               activeRepositoryClients.delete(repoPath);
             }
-            localDisposables.dispose();
-          };
 
-          // Allow consumers to use `onDidDestroy` for the ProjectRepos.
+            localDisposables.dispose();
+          }; // Allow consumers to use `onDidDestroy` for the ProjectRepos.
           // `getRootRepo()` can be used to add an onDidDestroy for the base
           // repo
-          this.onDidDestroy = function(callback: () => mixed): IDisposable {
+
+
+          this.onDidDestroy = function (callback) {
             localDisposables.add(callback);
             return {
               dispose() {
                 localDisposables.remove(callback);
-              },
+              }
+
             };
           };
         }
 
-        ProjectHgRepositoryClient.prototype = activeRepoClientInfo.repo;
+        ProjectHgRepositoryClient.prototype = activeRepoClientInfo.repo; // $FlowFixMe: this object has an HgRepositoryClient instance in its prototype chain
 
-        // $FlowFixMe: this object has an HgRepositoryClient instance in its prototype chain
-        return ((new ProjectHgRepositoryClient(): any): HgRepositoryClient);
+        return new ProjectHgRepositoryClient();
       } catch (err) {
-        logger.error(
-          'Failed to create an HgRepositoryClient for ',
-          directory.getPath(),
-          ', error: ',
-          err,
-        );
+        logger.error('Failed to create an HgRepositoryClient for ', directory.getPath(), ', error: ', err);
         return null;
       }
     });
   }
+
 }
+
+exports.default = HgRepositoryProvider;
